Make everything use int arrays

+++ src/DataSorter.java
@@ -2,5 +2,5 @@
  * Created by George on 2017-03-04.
  */
 public interface DataSorter {
-    void sort(String inputFile, String outputFile, int failureRate);
+    void sort(String inputFile, String outputFile, double failureRate);
 }
